move to asset config
