+++ runtime/selendra/src/config/mod.rs
@@ -1,3 +1,4 @@
+pub mod asset;
 pub mod bridge;
 pub mod consensus;
 pub mod dex;
